pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "health-engine"
version = "0.1.0"
description = "Health intelligence engine — scoring, insights, and wearable integrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
garmin = ["garminconnect>=0.2.0"]
dev = ["pytest>=7.0"]

[project.scripts]
health-engine = "cli:main"

[tool.setuptools.packages.find]
include = ["engine*"]
